Version 1.1

+++ webapp/install/sql/mysql_migrations/2012-07-19_checkin_support_v1.1.sql
@@ -7,4 +7,4 @@ ALTER TABLE tu_places ADD map_image VARCHAR( 255 ) COMMENT 'URL to an image of a
 --
 -- Add marker for whether or not a post archive has been loaded during a crawl.
 --
-ALTER TABLE tu_instances ADD  is_archive_loaded_posts INT( 1 ) NOT NULL DEFAULT  0 COMMENT  'Whether or not all the instance''s posts have been backfilled.' AFTER  earliest_reply_in_system;
+ALTER TABLE tu_instances ADD  is_archive_loaded_posts TINYINT( 1 ) NOT NULL DEFAULT  0 COMMENT  'Whether or not all the instance''s posts have been backfilled.' AFTER  earliest_reply_in_system;
